pyproject.toml
[project]
name = "autogui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "keyboard>=0.13.5",
    "mouse>=0.7.1",
    "opencv-python>=4.11.0.86",
    "paddleocr==3.3.1",
    "paddlepaddle==3.2.1",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "pydirectinput>=1.0.4",
    "pyside6>=6.11.0",
    "pywin32>=311",
    "setuptools>=78.1.0",
]

[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple"
